+++ tree/phylokernel.h
@@ -1519,7 +1519,7 @@ void PhyloTree::computePartialParsimonyFastSIMD(PhyloNeighbor *dad_branch, Phylo
                             }
                             x[state*VCSIZE + site/UINT_BITS] |= (1 << (site % UINT_BITS));
                         }
-                    } else if (state == (*alnit)->STATE_UNKNOWN) {
+                    } else if (state == static_cast<int>((*alnit)->STATE_UNKNOWN)) {
                         for (int j = 0; j < freq; j++, site++) {
                             if (site == NUM_BITS) {
                                 x += nstates*VCSIZE;
@@ -1562,7 +1562,7 @@ void PhyloTree::computePartialParsimonyFastSIMD(PhyloNeighbor *dad_branch, Phylo
                             }
                             x[state*VCSIZE + site/UINT_BITS] |= (1 << (site % UINT_BITS));
                         }
-                    } else if (state == (*alnit)->STATE_UNKNOWN) {
+                    } else if (state == static_cast<int>((*alnit)->STATE_UNKNOWN)) {
                         for (int j = 0; j < freq; j++, site++) {
                             if (site == NUM_BITS) {
                                 x += nstates*VCSIZE;
@@ -1596,7 +1596,7 @@ void PhyloTree::computePartialParsimonyFastSIMD(PhyloNeighbor *dad_branch, Phylo
                 int state = pat->at(leafid);
                 int freq = pat->frequency;
                 if (aln->seq_type == SEQ_POMO && state >= nstates 
-                    && state < aln->STATE_UNKNOWN) {
+                    && state < static_cast<int>(aln->STATE_UNKNOWN)) {
                     state -= nstates;
                     ASSERT(state < aln->pomo_sampled_states.size());
                     int id1 = aln->pomo_sampled_states[state] & 3;
@@ -1642,7 +1642,7 @@ void PhyloTree::computePartialParsimonyFastSIMD(PhyloNeighbor *dad_branch, Phylo
                         }
                         x[state*VCSIZE + site/UINT_BITS] |= (1 << (site % UINT_BITS));
                     }
-                } else if (state == (*alnit)->STATE_UNKNOWN) {
+                } else if (state == static_cast<int>((*alnit)->STATE_UNKNOWN)) {
                     for (int j = 0; j < freq; j++, site++) {
                         if (site == NUM_BITS) {
                             x += nstates*VCSIZE;
@@ -1662,7 +1662,7 @@ void PhyloTree::computePartialParsimonyFastSIMD(PhyloNeighbor *dad_branch, Phylo
             start_pos = end_pos;
         } // of end FOR LOOP
 
-        ASSERT(start_pos == aln->ordered_pattern.size());
+        ASSERT(static_cast<size_t>(start_pos) == aln->ordered_pattern.size());
 //        assert(site == aln->num_parsimony_sites % NUM_BITS);
         // add dummy states
         if (site > 0 && site < NUM_BITS) {
@@ -1797,7 +1797,7 @@ int PhyloTree::computeParsimonyBranchFastSIMD(PhyloNeighbor *dad_branch, PhyloNo
             VectorClass *x = (VectorClass*)(dad_branch->partial_pars + offset);
             VectorClass *y = (VectorClass*)(node_branch->partial_pars + offset);
             VectorClass w = x[0] & y[0];
-            for (size_t i = 1; i < nstates; i++) {
+            for (size_t i = 1; static_cast<int>(i) < nstates; i++) {
                 w |= x[i] & y[i];
             }
             w = ~w;
@@ -1882,7 +1882,7 @@ void PhyloTree::computePartialParsimonySankoffSIMD(PhyloNeighbor *dad_branch, Ph
                             *tip_buffer_ptr = partial_pars_child_ptr[j];
                     }
                     
-                    for (size_t i = 0; i < nstates; i++){
+                    for (size_t i = 0; static_cast<int>(i) < nstates; i++){
                         partial_pars_ptr[i] += tip_buffer[i];
                     }
                 } else {
@@ -1893,7 +1893,7 @@ void PhyloTree::computePartialParsimonySankoffSIMD(PhyloNeighbor *dad_branch, Ph
                     for (int i = 0; i < nstates; i++){
                         // min(j->i) from child_branch
                         VectorClass min_child_ptn_pars = partial_pars_child_ptr[0] + cost_matrix_ptr[0];
-                        for (size_t j = 1; j < nstates; j++) {
+                        for (size_t j = 1; static_cast<int>(j) < nstates; j++) {
                             min_child_ptn_pars = min(partial_pars_child_ptr[j] + cost_matrix_ptr[j], min_child_ptn_pars);
                         }
                         partial_pars_ptr[i] += min_child_ptn_pars;
@@ -1926,7 +1926,7 @@ void PhyloTree::computePartialParsimonySankoffSIMD(PhyloNeighbor *dad_branch, Ph
             }
             VectorClass *partial_pars_ptr = (VectorClass*)&partial_pars[ptn_start_index];
             
-            for (size_t i = 0; i < nstates; i++){
+            for (size_t i = 0; static_cast<int>(i) < nstates; i++){
                 // min(j->i) from child_branch
                 partial_pars_ptr[i] = tip_buffer[i] + tip_buffer_right[i];
             }
@@ -1952,10 +1952,10 @@ void PhyloTree::computePartialParsimonySankoffSIMD(PhyloNeighbor *dad_branch, Ph
             UINT *cost_matrix_ptr = cost_matrix;
             VectorClass right_contrib;
             
-            for(size_t i = 0; i < nstates; i++){
+            for(size_t i = 0; static_cast<int>(i) < nstates; i++){
                 // min(j->i) from child_branch
                 right_contrib = right_ptr[0] + cost_matrix_ptr[0];
-                for(size_t j = 1; j < nstates; j++) {
+                for(size_t j = 1; static_cast<int>(j) < nstates; j++) {
                     right_contrib = min(right_ptr[j] + cost_matrix_ptr[j], right_contrib);
                 }
                 partial_pars_ptr[i] = tip_buffer[i] + right_contrib;
@@ -1979,7 +1979,7 @@ void PhyloTree::computePartialParsimonySankoffSIMD(PhyloNeighbor *dad_branch, Ph
                 // min(j->i) from child_branch
                 left_contrib = left_ptr[0] + cost_matrix_ptr[0];
                 right_contrib = right_ptr[0] + cost_matrix_ptr[0];
-                for (size_t j = 1; j < nstates; j++) {
+                for (size_t j = 1; static_cast<int>(j) < nstates; j++) {
                     left_contrib = min(left_ptr[j] + cost_matrix_ptr[j], left_contrib);
                     right_contrib = min(right_ptr[j] + cost_matrix_ptr[j], right_contrib);
                 }
@@ -2046,7 +2046,7 @@ int PhyloTree::computeParsimonyBranchSankoffSIMD(PhyloNeighbor *dad_branch, Phyl
             VectorClass *dad_branch_ptr = (VectorClass*)&dad_branch->partial_pars[ptn_start_index];
             VectorClass min_ptn_pars = tip_buffer[0] + dad_branch_ptr[0];
             VectorClass br_ptn_pars = tip_buffer[0];
-            for (size_t i = 1; i < nstates; i++){
+            for (size_t i = 1; static_cast<int>(i) < nstates; i++){
                 // min(j->i) from node_branch
                 VectorClass min_score = tip_buffer[i] + dad_branch_ptr[i];
                 br_ptn_pars = select(min_score < min_ptn_pars, tip_buffer[i], br_ptn_pars);
@@ -2070,7 +2070,7 @@ int PhyloTree::computeParsimonyBranchSankoffSIMD(PhyloNeighbor *dad_branch, Phyl
                 // min(j->i) from node_branch
                 VectorClass min_score = node_branch_ptr[0] + cost_matrix_ptr[0];
                 VectorClass branch_score = cost_matrix_ptr[0];
-                for(size_t j = 1; j < nstates; j++) {
+                for(size_t j = 1; static_cast<int>(j) < nstates; j++) {
                     VectorClass value = node_branch_ptr[j] + cost_matrix_ptr[j];
                     branch_score = select(value < min_score, cost_matrix_ptr[j], branch_score);
                     min_score = min(value, min_score);
